initial commit
An async implementation of a segment structured tiered cache that
takes inspiration from both the Segcache and S3-FIFO papers.

This is a library that can be used as an embedded key-value cache.
It uses small FIFO queue for admission and either admits the item
to the cache or marks the hashtable entry as a ghost. Ghost items
continue to have frequency tracked so they can be admitted to the
main cache directly if they prove valuable.

The main cache uses a merge eviction strategy to retain high-value
items and either evicts or demotes lower value items. Demoted items
are moved to an on-disk admission queue and will either be admitted
to the main disk cache or marked as ghosts depending on if they are
accessed again while in the admission queue.

On access, items are promoted from disk back to the admission queue
for the in-memory cache. This avoids the need for additional
buffers to retrieve items from disk.